+++ packages/brick-container/src/i18n/constants.ts
@@ -1,0 +1,7 @@
+export const NS_BRICK_CONTAINER = "brick-container";
+
+export enum K {
+  BOOTSTRAP_ERROR = "BOOTSTRAP_ERROR",
+}
+
+export type Locale = { [key in K]: string };