+++ packages/brick-utils/src/placeholder/index.ts
@@ -3,4 +3,7 @@ import { utils } from "@easyops-cn/brick-next-pipes";
 export { inject, transform } from "./compile";
 
 // Keep compatibility of these exposed APIs.
-export const { formatUnitValue: formatValue, convertUnitValueByPrecision: convertValueByPrecision } = utils;
+export const {
+  formatUnitValue: formatValue,
+  convertUnitValueByPrecision: convertValueByPrecision,
+} = utils;