+++ packages/build-config-factory/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@2.0.0-next.7...@easyops/build-config-factory@2.0.0) (2020-09-13)
+
+**Note:** Version bump only for package @easyops/build-config-factory
+
+# [2.0.0-next.7](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@1.18.12...@easyops/build-config-factory@2.0.0-next.7) (2020-09-13)
+
+**Note:** Version bump only for package @easyops/build-config-factory
+
+# [2.0.0-next.6](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@1.18.11...@easyops/build-config-factory@2.0.0-next.6) (2020-09-04)
+
+**Note:** Version bump only for package @easyops/build-config-factory
+
+# [2.0.0-next.5](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@2.0.0-next.4...@easyops/build-config-factory@2.0.0-next.5) (2020-08-28)
+
+**Note:** Version bump only for package @easyops/build-config-factory
+
+# [2.0.0-next.4](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@1.18.9...@easyops/build-config-factory@2.0.0-next.4) (2020-08-28)
+
+**Note:** Version bump only for package @easyops/build-config-factory
+
+# [2.0.0-next.3](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@1.18.8...@easyops/build-config-factory@2.0.0-next.3) (2020-08-25)
+
+**Note:** Version bump only for package @easyops/build-config-factory
+
+# [2.0.0-next.2](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@1.18.7...@easyops/build-config-factory@2.0.0-next.2) (2020-08-13)
+
+**Note:** Version bump only for package @easyops/build-config-factory
+
+# [2.0.0-next.1](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@1.18.6...@easyops/build-config-factory@2.0.0-next.1) (2020-08-11)
+
+**Note:** Version bump only for package @easyops/build-config-factory
+
+# [2.0.0-next.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@1.18.2...@easyops/build-config-factory@2.0.0-next.0) (2020-08-05)
+
+**Note:** Version bump only for package @easyops/build-config-factory
+
+## [1.18.12](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@1.18.11...@easyops/build-config-factory@1.18.12) (2020-09-11)
+
+**Note:** Version bump only for package @easyops/build-config-factory
+
 ## [1.18.11](https://git.easyops.local/anyclouds/next-core/compare/@easyops/build-config-factory@1.18.10...@easyops/build-config-factory@1.18.11) (2020-09-02)
 
 **Note:** Version bump only for package @easyops/build-config-factory