+++ packages/dll-differ/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/dll-differ@1.0.0-next.3...@easyops/dll-differ@1.0.0) (2020-09-13)
+
+**Note:** Version bump only for package @easyops/dll-differ
+
+# [1.0.0-next.3](https://git.easyops.local/anyclouds/next-core/compare/@easyops/dll-differ@1.0.0-next.1...@easyops/dll-differ@1.0.0-next.3) (2020-08-28)
+
+### Bug Fixes
+
+- update brick-dll manifest.snapshot.json ([a258079](https://git.easyops.local/anyclouds/next-core/commits/a258079))
+
+# [1.0.0-next.1](https://git.easyops.local/anyclouds/next-core/compare/@easyops/dll-differ@0.0.1...@easyops/dll-differ@1.0.0-next.1) (2020-08-28)
+
+**Note:** Version bump only for package @easyops/dll-differ
+
 ## 0.0.1 (2020-08-26)
 
 **Note:** Version bump only for package @easyops/dll-differ