+++ packages/eslint-config-next/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/eslint-config-next@1.0.0-next.1...@easyops/eslint-config-next@1.0.0) (2020-09-13)
+
+**Note:** Version bump only for package @easyops/eslint-config-next
+
+# [1.0.0-next.1](https://git.easyops.local/anyclouds/next-core/compare/@easyops/eslint-config-next@0.1.18...@easyops/eslint-config-next@1.0.0-next.1) (2020-08-25)
+
+**Note:** Version bump only for package @easyops/eslint-config-next
+
 ## [0.1.18](https://git.easyops.local/anyclouds/next-core/compare/@easyops/eslint-config-next@0.1.17...@easyops/eslint-config-next@0.1.18) (2020-08-23)
 
 **Note:** Version bump only for package @easyops/eslint-config-next